feat(showComment) : 댓글 조회 기능 추가, 대댓글 삭제 기능 추가
댓글 조회 기능 구현, 부모 댓글 삭제 시 대댓글도 삭제되도록 벌크 연산 이용

+++ src/main/java/com/yht/exerciseassist/domain/post/dto/PostListDto.java
@@ -15,11 +15,11 @@ public class PostListDto {
     private int mediaListCount;
     private int likeCount;
     private Long views;
-    private Long commentCount;
+    private int commentCount;
 
     @Builder
     public PostListDto(String username, String postType, String workOutCategory, String createdAt,
-                       String title, int mediaListCount, int likeCount, Long views, Long commentCount, Long postId) {
+                       String title, int mediaListCount, int likeCount, Long views, int commentCount, Long postId) {
         this.username = username;
         this.postType = postType;
         this.workOutCategory = workOutCategory;
